fix for default y-axis
Changed arc marker to be a circle marker. This secondary marker is smaller than the first one so you can see both together when a point is both concern and improvement. This has fixed the 0 y-axis default issue. You can now determine if you want to fix to 0 or not with the parameter PinYAxisToZero.